chore: move vllm_engine to model/velocity
