+++ ChatTTS/model/velocity/model_loader.py
@@ -22,7 +22,7 @@ def _set_default_torch_dtype(dtype: torch.dtype):
 
 
 def _get_model_architecture(config: PretrainedConfig) -> Type[nn.Module]:
-    model_cls = getattr(importlib.import_module("ChatTTS.vllm_engine.llama"), "LlamaModel", None)
+    model_cls = getattr(importlib.import_module("ChatTTS.model.velocity.llama"), "LlamaModel", None)
     return model_cls
 
 def get_model(model_config: ModelConfig) -> nn.Module:
