+++ ChatTTS/model/velocity/worker.py
@@ -14,7 +14,7 @@ from vllm.model_executor.parallel_utils.parallel_state import (
     initialize_model_parallel)
 from vllm.sequence import SamplerOutput, SequenceGroupMetadata
 from vllm.worker.cache_engine import CacheEngine
-from ChatTTS.vllm_engine.model_runner import ModelRunner
+from ChatTTS.model.velocity.model_runner import ModelRunner
 
 
 class Worker:
